flask-client/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
contextlib2==0.6.0.post1
Flask==1.1.1
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pystardog==0.9.6
requests==2.22.0
requests-toolbelt==0.9.1
urllib3==1.25.7
Werkzeug==0.16.0
